requirements.txt
flask==3.1.3
tensorflow==2.21.0
pillow==12.1.1
numpy==2.4.4
gunicorn==21.2.0